+++ app/src/main/java/pk/sufiishq/app/feature/app/SyncManager.kt
@@ -26,6 +26,8 @@ import pk.sufiishq.app.feature.events.worker.EventSyncWorker
 import pk.sufiishq.app.feature.events.worker.EventUpdateWorker
 import pk.sufiishq.app.feature.occasions.worker.OccasionSyncWorker
 import pk.sufiishq.app.feature.storage.KeyValueStorage
+import pk.sufiishq.app.feature.theme.controller.ThemeController
+import pk.sufiishq.app.feature.theme.model.AutoChangeColorDuration
 import pk.sufiishq.app.worker.CacheRemoveWorker
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -38,12 +40,19 @@ class SyncManager @Inject constructor(
     @SharedPreferences private val keyValueStorage: KeyValueStorage,
 ) {
 
-    fun sync() {
+    fun sync(themeController: ThemeController) {
         CoroutineScope(dispatcher).launch {
             CacheRemoveWorker.init(context)
             EventUpdateWorker.init(context)
             EventSyncWorker.init(context, keyValueStorage)
             OccasionSyncWorker.init(context, keyValueStorage)
+            initAutoColorChangeWorker(themeController)
+        }
+    }
+
+    private suspend fun initAutoColorChangeWorker(themeController: ThemeController) {
+        if (!themeController.isAutoChangeColorEnable() && themeController.getActiveAutoColorChangeDuration() == null) {
+            themeController.setAutoChangeColor(true, AutoChangeColorDuration.every1Hour())
         }
     }
 }