+++ app/src/main/java/pk/sufiishq/app/feature/theme/controller/ThemeController.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2022-2023 SufiIshq
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package pk.sufiishq.app.feature.theme.controller
 
 import pk.sufiishq.app.feature.theme.model.AutoChangeColorDuration
@@ -7,7 +23,7 @@ interface ThemeController {
 
     suspend fun isAutoChangeColorEnable(): Boolean
     fun setAutoChangeColor(isEnable: Boolean, activeDuration: AutoChangeColorDuration)
-    suspend fun getActiveAutoColorChangeDuration(): AutoChangeColorDuration
+    suspend fun getActiveAutoColorChangeDuration(): AutoChangeColorDuration?
     suspend fun getAutoColorChangeDurationList(): List<AutoChangeColorDuration>
     fun updateAutoColorChangeDuration(autoChangeColorDuration: AutoChangeColorDuration)
     suspend fun getAvailableColorPalettes(): List<ColorPalette>